Do not supply cutting with timestamp
The timestamp should belong to the first character byte received, not to
the character end. This makes implementation simpler and seems to be
more appropriate security-wise.

After all even though we can't log a non-complete character unless
forced, the software on the either end might react to it at the time it
was transferred and we should record *that* time.