Add session locking support
Create a per-session lock file to prevent duplicate recording of
sessions. The dedicated lock directory is assumed to be accessible only
to tlog-rec SUID/SGID IDs.

When certain failures occur while creating a lock file, assume session
was unlocked, and record anyway, as it is safer to record a session than
not. Ignore all errors removing a lock file.

Fixes #4