Handle piped in I/O from STDIN
Modify the main recording loop to close corresponding I/O sink
file descriptors when EOF is signalled from a source read. This
is necessary to close one end of a pipe when I/O is pushed
to an interactive process like python via stdin.

In addition, execute the main recording transfer loop until the
child exits and a EIO/EOF condition is met from a source read.