issue 283: tlog-play improve auth options.
As well as being able to specify authentication in the URL, provide some other
options for configuring the curl client that connects to the ElasticSearch API.

* user:pass in the TLOG_CURL_USERPWD environment variable
* In a ~/.netrc file
* Enable all curl auth methods, so kerberos is now possible.

For kerberos the username must be non-empty, either in the URL or $TLOG_CURL_USERPWD.

Make ES file reader fail early and with better error, if HTTP response >= 400.

Add --es-verbose option for verbose logging from curl.