Correct suse rpmbuild with included spec file
The package name in SuSE for the utempter devel package differs
from redhat.  Also, the default documentation directory differs
between these causing the configure script to choose a different
location than the %files section assumes.  This commit makes
minor adjustments to the build to allow rpmbuild to work on SuSE
out of the box.