tlitest: extend delay for limit-action delay test
In some environments the limit-action delay test listed below is
failing:

test_record_fast_input_with_limit_action_delay

This requires more time to write to the output log file set by
tlog-rec in the test.  In order to accomplish this, the check_outfile
misc function is modified to take a parameter called maxchecks which can
be used to set how many times to check for a string in the file in
between the 5 second delays.

For the limit-action delay test, this is set to 100 to wait long enough
to catch the last message being written to the log file.